engine: add engine config builder
Add a configuration and builder for the Hop engine. The builder
currently supports setting whether to enable pubsub and the maximum
number of active sessions.

Closes #5.
